README
5. Performance

The PubSub tutorial instructs you to create a consumer program that calls print() for each record. This is good for debugging, but it can lead to performance issues. Next, you will measure the performance effect of the print() calls.

Create a new data file that contains one day of data for all of the Vehicle IDs assigned to you. 

Run your producer to send all of this data to your PubSub topic. How long does it take for the producer to send all of the data?

Run your consumer to consume all of the data. How long does it take for the consumer to consume all of the data?

Remove the line print(f"Received {message}.") from your consumer and add code to the consumer to count the number of messages received and to print this count once at the very end.  How many breadcrumbs are produced and consumed?

Rerun part B. How long does it take to run?

Rerun part C using the new consumer. How long does it take to run?

Did changing the consumer affect the time to run?  How much?